added tow new opponents and modified some others
