+++ GameLogic/src/test/java/opponent/AllOpponentsTest.java
@@ -24,6 +24,8 @@ public class AllOpponentsTest {
     @BeforeEach
     void init() {
         opponents = new Opponent[]{
+                new TonyRandomOpponent(),
+                new OleOpponent(),
                 new RandomOpponent(),
         };
     }
